% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resymbol.R
\name{resymbol}
\alias{resymbol}
\title{Update Gene Symbols}
\usage{
resymbol(hgnc_complete_set = hgnc_complete_set, nameold)
}
\arguments{
\item{hgnc_complete_set}{A dataframe object import from file which can be downloaded from HGNC genenamesorg archived datasets at https://ftp.ebi.ac.uk/pub/databases/genenames/hgnc/archive/.}

\item{nameold}{A vector of symbols needed to be checked}
}
\value{
A dataframe including replaced symbols. old: symbols input. newnames: returned new symbols. prev/aliaschange: newnames searched by previous/alias symbol list. changenumber: how many new symbols did the old symbol matched. newname_to_use: unique updated symbol list
}
\description{
Using local gene files to update symbols. Only unique symbols will be returned.
}
\details{
Old symbols --> symbols not included in the newest symbol list --> symbols that matches only one time with previous/alias symbol --> matched new symbols for the remaining. The input symbols that didn't match the new symbol column in hgnc_complete_set.txt will be regarded as old symbols. Old symbols that matches only one time in first previous symbols list and if no matches then the alias symbols list will be used for matching new symbols.
}
\examples{
hgnc_complete_set <- read.delim("hgnc_complete_set.txt")
nameold = c("TP53","CD206")
nametotal = resymbol(hgnc_complete_set,nameold)
}
\author{
Tu Jihao
}
